move poseidon hasher and merger to separate files
