+++ merkle/src/lib.rs
@@ -17,8 +17,12 @@ mod masking;
 pub use masking::*;
 mod merger;
 pub use merger::*;
+mod merger_poseidon;
+pub use merger_poseidon::*;
 mod hasher;
 pub use hasher::*;
+mod hasher_poseidon;
+pub use hasher_poseidon::*;
 mod metadata;
 pub use metadata::*;
 pub mod prefixes;
